+++ malachite-nz/src/bin_util/demo_and_bench/gaussian_integer/arithmetic/mod.rs
@@ -12,8 +12,10 @@ pub(crate) fn register(runner: &mut Runner) {
     abs_squared::register(runner);
     add::register(runner);
     conjugate::register(runner);
+    div_i::register(runner);
     is_power_of_2::register(runner);
     mul::register(runner);
+    mul_i::register(runner);
     neg::register(runner);
     power_of_2::register(runner);
     shl::register(runner);
@@ -24,8 +26,10 @@ pub(crate) fn register(runner: &mut Runner) {
 mod abs_squared;
 mod add;
 mod conjugate;
+mod div_i;
 mod is_power_of_2;
 mod mul;
+mod mul_i;
 mod neg;
 mod power_of_2;
 mod shl;